# src/main/webapp/docs/swagger.yaml
swagger: '2.0'
info:
  description: Api para buscar contenido audiovisual
  version: 1.0.0
  title: MovieFinder
  # put the contact info for your development or API team
  contact:
    email: alvagualh@alum.us.es
servers:
  url: https://snmanager.appspot.com/api
  description: Entrada a la api desplegada
# tags are used for organizing operations
...
paths:
...
  /actor:
    get:
      tags: 
      - Actor
      summary: getAllActores
      operationId: getAllActores
      description: Devuelve todos los actores
      produces: 
      - application/json
      responses:
        200:
          description: Devolviendo contenido
          schema:
            $ref: '#/definitions/Actor'
    post:
      tags:
      - Actor
      summary: addActor
      parameters:
      - in: body
        name: addActor
        description: Añadir actor
        schema:
          $ref: '#/definitions/Actor'
      operationId: addActor
      description: Añadir un actor
      produces:
      - application/json
      responses:
        201:
          description: Created
    put:
      tags:
      - Actor
      summary: updateActor
      parameters:
      - in: body
        name: updateActor
        description: Actualizar actor
        schema:
          $ref: '#/definitions/Actor'
      operationId: updateActor
      description: Actualizar actor
      produces:
      - application/json
      responses:
        204:
          description: No contend
        404:
          description: Not found
  /actor/{actorId}:
    get:
      tags:
      - Actor
      summary: getActor
      parameters:
      - in: path
        name: actorId
        required: true
        description: Id de actor
        schema:
          type: string
      operationId: getActor
      description: Devuelve un actor
      produces:
      - application/json
      responses:
        200:
          description: Devuelve el actor
          schema:
            $ref: '#/definitions/Actor'
    delete:
      tags:
      - Actor
      summary: deleteActor
      parameters:
      - in: path
        name: actorId
        required: true
        description: Id de actor
        schema:
          type: string
      operationId: deleteActor
      description: Elimina un actor
      produces:
      - application/json
      responses:
        204:
          description: No content
        404:
          description: Not found
...
  /tipo:
    post:
      tags:
      - Tipo
      summary: addTipo
      parameters:
      - in: body
        name: addTipo
        description: Añadir tipo
        schema:
          $ref: '#/definitions/Tipo'
      operationId: addTipo
      description: Añadir un tipo
      produces:
      - application/json
      responses:
        201:
          description: Created
  /tipo/{tipoId}:
    get:
      tags:
      - Tipo
      summary: getTipo
      parameters:
      - in: path
        name: tipoId
        required: true
        description: Id de tipo
        schema:
          type: string
      operationId: getTipo
      description: Devuelve un tipo
      produces:
      - application/json
      responses:
        200:
          description: Devuelve el tipo
          schema:
            $ref: '#/definitions/Tipo'
...
  /audiovisual:
    get:
      tags: 
      - Audiovisual
      summary: getAllAudiovisuales
      operationId: getAllAudiovisuales
      description: Devuelve todas los audiovisuales
      produces: 
      - application/json
      responses:
        200:
          description: Devolviendo contenido
          schema:
            $ref: '#/definitions/AudioVisual'
    post:
      tags:
      - Audiovisual
      summary: addAudiovisual
      parameters:
      - in: body
        name: addAudiovisual
        description: Añadir audiovisual
        schema:
          $ref: '#/definitions/AudioVisual'
      operationId: addAudiovisual
      description: Añadir un audiovisual
      produces:
      - application/json
      responses:
        201:
          description: Created
    put:
      tags:
      - Audiovisual
      summary: updateAudiovisual
      parameters:
      - in: body
        name: updateAudiovisual
        description: Actualizar audiovisual
        schema:
          $ref: '#/definitions/AudioVisual'
      operationId: updateAudiovisual
      description: Actualizar audiovisual
      produces:
      - application/json
      responses:
        204:
          description: No contend
        404:
          description: Not found
  /audiovisual/{audiovisualId}:
    get:
      tags:
      - Audiovisual
      summary: getAudiovisual
      parameters:
      - in: path
        name: audiovisualId
        required: true
        description: Id del audiovisual
        schema:
          type: string
      operationId: getAudiovisual
      description: Devuelve un audiovisual
      produces:
      - application/json
      responses:
        200:
          description: Devuelve el audiovisual
          schema:
            $ref: '#/definitions/AudioVisual'
    delete:
      tags:
      - Audiovisual
      summary: deleteAudiovisual
      parameters:
      - in: path
        name: audiovisualId
        required: true
        description: Id del audiovisual
        schema:
          type: string
      operationId: deleteAudiovisual
      description: Elimina un audiovisual
      produces:
      - application/json
      responses:
        204:
          description: No content
        404:
          description: Not found
  /audiovisual/{audiovisualId}/{playlistId}:
    post:
      tags:
      - Audiovisual
      summary: addPlaylistAAudiovisual
      parameters:
      - in: path
        name: audiovisualId
        required: true
        schema:
          type: string
        description: Id del audiovisual
      - in: path
        name: playlistId
        required: true
        schema:
          type: string
        description: Id de la playlist
      operationId: addPlaylistAAudiovisual
      description: Añade playlist a audiovisual
      produces:
      - application/json
      responses:
        201:
          description: Created
        400:
          description: Bad request
        404:
          description: Not found
    delete:
      tags:
      - Audiovisual
      summary: deletePlaylistDeAudiovisual
      parameters:
      - in: path
        name: audiovisualId
        required: true
        schema:
          type: string
        description: Id del audiovisual
      - in: path
        name: playlistId
        required: true
        schema:
          type: string
      operationId: deletePlaylistDeAudiovisual
      description: Elimina playlist del audiovisual
      produces:
      - application/json
      responses:
        204:
          description: No contend
        404:
          description: Not found
  /audiovisual/{audiovisualId}/{tipoId}:
    post:
      tags:
      - Audiovisual
      summary: addTipoAAudiovisual
      parameters:
      - in: path
        name: audiovisualId
        required: true
        schema:
          type: string
        description: Id del audiovisual
      - in: path
        name: tipoId
        required: true
        schema:
          type: string
        description: Id del tipo
      operationId: addTipoAAudiovisual
      description: Añade tipo a audiovisual
      produces:
      - application/json
      responses:
        201:
          description: Created
        400:
          description: Bad request
        404:
          description: Not found
    delete:
      tags:
      - Audiovisual
      summary: deleteTipoDeAudiovisual
      parameters:
      - in: path
        name: audiovisualId
        required: true
        schema:
          type: string
        description: Id del audiovisual
      - in: path
        name: tipoId
        required: true
        schema:
          type: string
      operationId: deleteTipoDeAudiovisual
      description: Elimina tipo del audiovisual
      produces:
      - application/json
      responses:
        204:
          description: No contend
        404:
          description: Not found
  /audiovisual/{audiovisualId}/{actorId}:
    post:
      tags:
      - Audiovisual
      summary: addActorAAudiovisual
      parameters:
      - in: path
        name: audiovisualId
        required: true
        schema:
          type: string
        description: Id del audiovisual
      - in: path
        name: actorId
        required: true
        schema:
          type: string
        description: Id del actor
      operationId: addActorAAudiovisual
      description: Añade actor a audiovisual
      produces:
      - application/json
      responses:
        201:
          description: Created
        400:
          description: Bad request
        404:
          description: Not found
    delete:
      tags:
      - Audiovisual
      summary: deleteActorDeAudiovisual
      parameters:
      - in: path
        name: audiovisualId
        required: true
        schema:
          type: string
        description: Id del audiovisual
      - in: path
        name: actorId
        required: true
        schema:
          type: string
      operationId: deleteActorDeAudiovisual
      description: Elimina actor del audiovisual
      produces:
      - application/json
      responses:
        204:
          description: No contend
        404:
          description: Not found        
definitions:
  AudioVisual:
    type: object
    required:
    - id
    - nombre
    properties:
      id:
        type: string
        example: av1
      nombre:
        type: string
        example: Bohemian Rhapsody
      tipo:
        $ref: '#/definitions/Tipo'
      actores:
        type: array
        items:
          $ref: '#/definitions/Actor'
      soundtrack:
        $ref: '#/definitions/Playlist'
  Actor:
    required:
    - id
    - nombre
    properties:
      id:
        type: string
        example: a1
      nombre:
        type: string
        example: Rami Malek
      nacionalidad:
        type: string
        example:  Estadounidense
      fechaNacimiento:
        type: string
        format: date
        example: 17-10-1980
      fechaMuerte:
        type: string
        format: date
        example: 17-10-2020
      premios:
        type: string
        example: Emmy, oscar
  Playlist:
    required:
    - id
    - nombre
    properties:
      id:
        type: string
        example: p1
      nombre:
        type: string
        example: Bohemian Rhapsody ST
      description:
        type: string
        example: Banda sonora de la pelicula Bohemian Rhapsody
      songs:
        type: array
        items:
          $ref: '#/definitions/Song'
  Song:
    required:
    - id
    - title
    properties:
      id:
        type: string
        example: so1
      title:
        type: string
        example: Show must go on
      artist:
        type: string
        example: Queen
      album:
        type: string
        example: Bohemian Rhapsody
      year:
        type: string
        example: 1970
  Tipo:
    required:
    - id
    properties:
      id:
        type: string
        example: t1
      serie:
        $ref: '#/definitions/Serie'
      pelicula:
        $ref: '#/definitions/Pelicula'
  Pelicula:
    required:
    - id
    - director
    properties:
      id:
        type: string
        example: pe1
      director:
        type: string
        example: Cristopher Nolan
      fechaEstreno:
        type: string
        format: date
        example: 15-11-2019
  Serie:
    required:
    - id
    - temporadas
    properties:
      id:
        type: string
        example: s1
      temporadas:
        type: integer
        example: 2
      ultFechaEmision:
        type: string
        format: date
        example: 20-10-2017
      estreno:
        type: string
        format: date
        example: 12-9-2013
      canal:
        type: string
        example: HBO
# Added by API Auto Mocking Plugin
# host: snmanager.appspot.com
# basePath: /api/
schemes:
 - https
# Added by API Auto Mocking Plugin
host: virtserver.swaggerhub.com
basePath: /alvagualh/MovieFinder/1.0.0